+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   the involved dictionary fields' values have changed
 * [HexaPDF::Type::AcroForm::ChoiceField#list_box_top_index=] to raise an error
   if no option items are set
+* [HexaPDF::PDFArray#to_ary] to return an array with preprocessed values
 
 
 ## 0.13.0 - 2020-11-15